+++ lib/poolex/state.ex
@@ -1,25 +1,21 @@
 defmodule Poolex.State do
   @moduledoc false
-  defstruct busy_workers_count: 0,
-            busy_workers_pids: [],
-            idle_workers_count: 0,
-            idle_workers_pids: [],
+  defstruct busy_workers_state: nil,
+            idle_workers_state: nil,
+            waiting_callers_state: nil,
             worker_module: nil,
             worker_start_fun: nil,
             worker_args: [],
-            waiting_callers: :queue.new(),
             monitor_id: nil,
             supervisor: nil
 
   @type t() :: %__MODULE__{
-          busy_workers_count: non_neg_integer(),
-          busy_workers_pids: list(pid()),
-          idle_workers_count: non_neg_integer(),
-          idle_workers_pids: list(pid()),
+          busy_workers_state: Poolex.Workers.Behaviour.state(),
+          idle_workers_state: Poolex.Workers.Behaviour.state(),
+          waiting_callers_state: Poolex.Callers.Behaviour.state(),
           worker_module: module(),
           worker_start_fun: atom(),
           worker_args: list(any()),
-          waiting_callers: :queue.queue(),
           monitor_id: atom() | reference(),
           supervisor: pid()
         }